+++ src/Toolkit/tests/Command/CreateKitCommandTest.php
@@ -88,7 +88,12 @@ class CreateKitCommandTest extends KernelTestCase
         $this->assertStringEqualsFile(
             $this->tmpDir.'/Button/templates/components/Button.html.twig',
             <<<'TWIG'
-                {% props type = 'button', variant = 'default' %}
+                {% props
+                    ## 'button'|'submit'|'reset' The button type.
+                    type = 'button',
+                    ## 'default'|'secondary' The visual style variant.
+                    variant = 'default'
+                %}
                 {%- set style = html_cva(
                     base: 'inline-flex items-center',
                     variants: {
@@ -105,6 +110,7 @@ class CreateKitCommandTest extends KernelTestCase
                         type: 'submit',
                     }) }}
                 >
+                    {## The button label. #}
                     {%- block content %}{% endblock -%}
                 </button>
                 TWIG